Add files via upload
Vital Mind is a platform designed to provide users with a diverse range of courses in categories such as Technology, Spirituality, Health and Fitness, Business, and Juniors. Users can browse courses tailored to their needs, select expert mentors in their area of interest, and enroll in paid courses through a seamless online experience.